+++ src/static_folders/__init__.py
@@ -1,4 +1,4 @@
-from .folder import Folder, _FolderName, custom_name
-from .partitioned_folder import FolderPartition, EnumeratedFolderPartition
+from .folder import Folder
+from .partitioned_folder import EnumeratedFolderPartition, FolderPartition
 
-__all__ = ["EnumeratedFolderPartition", "Folder", "FolderPartition", "_FolderName", "custom_name"]
+__all__ = ["EnumeratedFolderPartition", "Folder", "FolderPartition"]