+++ src/static_folders/folder.py
@@ -4,10 +4,10 @@ import inspect
 import os
 import sys
 import typing
-from pathlib import Path, PureWindowsPath, PurePosixPath
-from typing import Sequence, Any, Callable, TypeVar, ClassVar, Type, overload
+from pathlib import Path
+from typing import Any, Callable, ClassVar, Sequence, Type, TypeVar
 
-from attrs import define, field, Factory
+from attrs import Factory, define, field
 from typing_extensions import Self
 
 from static_folders.folder_interface import FolderLike
@@ -21,38 +21,6 @@ U = TypeVar("U", bound="Folder")
 
 PathLike = typing.Union[str, Path]
 T = TypeVar("T", bound="Folder")
-
-V = TypeVar("V")
-
-if sys.platform == "win32":
-
-    class _FolderName(PureWindowsPath):
-        """TODO this subclassing might be a bad idea"""
-else:
-
-    class FolderName(PurePosixPath):
-        """TODO this subclassing might be a bad idea"""
-
-
-@overload
-def custom_name(name: str, annotation_type: None = ...) -> Folder: ...
-
-
-@overload
-def custom_name(name: str, annotation_type: Type[U] = ...) -> U: ...
-
-
-def custom_name(name: str, annotation_type: Type[U] | None = None) -> U | Folder:
-    # We are lying to the type system here, and assuming users don't use this
-    return _FolderName(name)  # type:ignore[return-value]
-
-
-# def name(value):
-#     """First draft value to distinguish folder directory overrides.
-#
-#     # TODO should we just have FolderName / FileName classes for consistency?
-#     """
-#     return CustomFolderName(value)
 
 
 def _get_annotations(obj: Callable[..., object] | type[Any] | ModuleType) -> dict[str, object]:
@@ -93,7 +61,7 @@ class Folder(FolderLike):
     _child_folders: list[FolderLike] = field(init=False, default=Factory(list))
 
     @classmethod
-    def from_path(cls, path: Path) -> Folder:
+    def from_path(cls, path: Path) -> Self:
         return cls(path)
 
     @classmethod
@@ -131,43 +99,44 @@ class Folder(FolderLike):
             )
             raise TypeError(msg)
 
+        # resolve all the child attributes from annotations into instance objects
         for attrib_name, annotation in annotations.items():
             if attrib_name in self._reserved_attributes:
                 continue
             if isinstance(annotation, type):
-                if issubclass(annotation, FolderLike):  # i.e. attribute foo: Folder - a class constructor
+                if issubclass(annotation, FolderLike):  # e.g. `foo: Folder`
                     value = getattr(self, attrib_name, None)
                     folder_name = None
                     if value is None:  # check default wasn't given
                         folder_name = attrib_name
-                    elif isinstance(value, _FolderName):
-                        folder_name = value.name
-                    elif isinstance(value, FolderLike):
+                    elif isinstance(value, FolderLike):  # `foo: Folder = Folder("bar")` -> foo / "bar"
+                        folder_name = value.location.name
+                    elif isinstance(value, Path):  # `foo: Folder = Path("bar")`
+                        # This is not permitted by the type system, but is an easy typo someone could make
+                        # which would lead to confusing behaviour - the path wouldn't be "seen" at all.
                         msg = (
-                            f"Providing a FolderLike annotation with a FolderLike value "
-                            f"({attrib_name}: {annotation} = {value}) is deprecated, "
-                            "behaviour was not sound with respect to child file paths. If the intention "
-                            "was to specify a custom folder name, you should "
-                            f"migrate to {attrib_name}: {annotation} = sf.custom_name(...) "
-                        )
-                        raise TypeError(msg)
-                    elif isinstance(value, Path):
-                        # TODO, if we're doing all this at runtime, do we get it at type checking time?
-                        # and if not, does that defeat the purpose of this being staticly typed.
-                        msg = (
-                            f"Annotating an attribute with a FolderLike type and a Path value is ambiguous "
-                            f"and not supported (got {attrib_name}: {annotation} = {value}). "
+                            f"Annotating an attribute with a FolderLike type and a Path value is incorrect "
+                            f"and not supported. For class `{type(self).__name__}` got:\n"
+                            f"'{attrib_name}: {annotation.__name__} = {value!r}'\n"
                             f"If you intended to declare a subfolder with a custom folder name, "
-                            f"you should update the Path value to a sf.custom_name(...) value instead.\n"
-                            f"If you intended to declare a file within the folder, you should update the "
-                            f"FolderLike annotation to be a Path instead."
+                            f"you should use `{attrib_name}: FolderLike = FolderLike(name)` instead.\n"
+                            f"If you intended to declare a file within the folder, you should use"
+                            f" `{attrib_name}: Path: Path(name)` "
+                            f"instead."
                         )
                         raise TypeError(msg)
                     else:
                         pass  # subclasses or lambda come through this passage
                         # print("pre-existing thing", value)
                     if folder_name is not None:
-                        value = annotation.from_path(self.location / folder_name)
+                        try:
+                            value = annotation.from_path(self.location / folder_name)
+                        except TypeError as e:
+                            msg = (
+                                f"Building Class {type(self).__name__} failed on constructing attribute:\n"
+                                f"'{attrib_name}: {annotation.__name__} = {value!r}' with attached error"
+                            )
+                            raise TypeError(msg) from e
                         setattr(self, attrib_name, value)
 
                         self._child_folders.append(value)
@@ -191,13 +160,16 @@ class Folder(FolderLike):
                         raise TypeError(msg)
                     setattr(self, attrib_name, self.location / provided_path)
                 elif issubclass(annotation, str):
+                    value = getattr(self, attrib_name, "")
                     msg = (
                         "Folder subclasses do not support raw str annotated fields "
                         f"to avoid ambiguity between whether the str represents a filepath or not. "
-                        f"Class {self.__class__} has attribute {attrib_name!r} "
-                        f"which is annotated as a string. If your intention is to provide a relative filepath, "
-                        f"provide a Path instead. If your intention is to declare a class variable string, "
-                        f"use the full ClassVar[str] annotation to convey this."
+                        f"Class {type(self).__name__!r} defines the attribute:\n"
+                        f"'{attrib_name}: {annotation.__name__} = {value!r}'\n"
+                        f"which is annotated as a string. If you meant to:\n"
+                        f"  - Specify a child file -> use a `Path` instead\n"
+                        f"  - Specify a child folder -> use a `FolderLike` instead\n"
+                        f"  - Specify a class variable string -> annotate with `ClassVar[str]` instead\n"
                     )
                     raise TypeError(msg)